Mini-stats - separated file into per-class files.
